Move staging db with data in shared script
