+++ support/setup/Gemfile
@@ -1,0 +1,5 @@
+source "https://rubygems.org"
+
+gem "pg"
+gem "mysql2"
+gem "activerecord"
